added cppunit tests for Event and BarEvent

+++ Event/BarEvent.cpp
@@ -20,7 +20,7 @@
  */
 BarEvent::BarEvent(std::string ticker_, std::string time_, int period_,
 		double openPrice_, double highPrice_, double lowPrice_,
-		double closePrice_, long int volume_, double adjClosePrice_):
+		double closePrice_, int volume_, double adjClosePrice_):
 				Event::Event("BAR"), m_ticker(ticker_), m_time(time_),
 				m_period(period_), m_openPrice(openPrice_), m_highPrice(highPrice_),
 				m_lowPrice(lowPrice_), m_closePrice(closePrice_), m_volume(volume_),
